:truck: Rename ActiveRecord::Dbt::Factory::ColumnsFactory to ActiveRecord::Dbt::Factory::Columns::YmlFactory
